nrfx_qspi: Correct use of workaround for anomaly 215 and 43 in nrfx_qspi
The code that accesses QSPI registers above 0x600 needs to apply
workaround for anomaly 215 on nRF52840 or anomaly 43 on nRF5340.

The commit removes check if a custom instruction long transfer
is ongoing from deactivate function. The check could trigger
anomalies and now it's user responsibility to be sure that custom
instruction long transfer has ended otherwise it will be interrupted.

Change affects behaviour of nrfx_qspi_uinit() and nrfx_qspi_deactivate()
what is mentioned in documentation for that functions.

Signed-off-by: Adam Wojasinski <adam.wojasinski@nordicsemi.no>